# app/definition/swagger.yaml
swagger: "2.0"
info:
  version: desp - v1
  title: Despachalo.com
  description: API de servicios de Despachalo.com
# during dev, should point to your local machine
host: localhost:10010
...
basePath: /desp/v1
...
schemes:
  # tip: remove http to make production-grade
  - http
  - https
# format of bodies a client can send (Content-Type)
consumes:
  - application/json
# format of the responses to the client (Accepts)
produces:
  - application/json
paths:
...
  /permission:
    get:
      tags:
        - Permisos
      description: Lista los permisos
      # used as the method name of the controller
      operationId: findPermissions
      parameters:
        - name: authorization
          in: header
          description: Json Web Token
          required: true
          type: string
        - name: tipo
          in: query
          description: ID Tipo de usuario
          required: false
          type: string
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/DespachaloResponse"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
    post:
      tags:
        - Permisos
      description: Registrar Permisos
      # used as the method name of the controller
      operationId: addPermission
      parameters:
        - name: authorization
          in: header
          description: Json Web Token
          required: true
          type: string
        - name: data
          in: body
          required: true
          schema:
            $ref: "#/definitions/Permission"
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/DespachaloResponse"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
    put:
      tags:
        - Permisos
      description: Editar Permisos
      # used as the method name of the controller
      operationId: permisosEditar
      parameters:
        - name: authorization
          in: header
          description: Json Web Token
          required: true
          type: string
        - name: data
          in: body
          required: true
          schema:
            $ref: "#/definitions/Permission"
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/DespachaloResponse"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
    delete:
      tags:
        - Permisos
      description: Elimina un Permiso
      # used as the method name of the controller
      operationId: permisosBorrar
      parameters:
        - name: authorization
          in: header
          description: Json Web Token
          required: true
          type: string
        - name: data
          in: body
          required: true
          schema:
            $ref: "#/definitions/PermissionDelete"
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/DespachaloResponse"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
...
  /puntopagos/pagar/:
    # binds a127 app logic to a route
    x-swagger-router-controller: botonPuntoPagos
    x-a127-apply: {}
    get:
      tags:
        - PuntoPago
      description: Recibe monto a recaudar por PuntoPagos
      # used as the method name of the controller
      operationId: puntoPagosPagar
      parameters:
        - name: monto
          in: query
          description: Monto a recaudar
          required: true
          type: string
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/DespachaloResponse"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /puntopagos/notificacion/:
    # binds a127 app logic to a route
    x-swagger-router-controller: botonPuntoPagos
    x-a127-apply: {}
    post:
      tags:
        - PuntoPago
      description: Punto Pagos notifica que se efectuó el pago
      # used as the method name of the controller
      operationId: puntoPagosNotificacion
      parameters:
        - name: fecha
          in: header
          description: Fecha según la especificación RFC1123
          required: true
          type: string
        - name: autorizacion
          in: header
          description: Firma del mensaje utilizando la llave secreta
          required: true
          type: string
        - name: data
          in: body
          required: true
          schema:
            $ref: "#/definitions/Notificacion"
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/DespachaloResponse"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /puntopagos/validar/:
    # binds a127 app logic to a route
    x-swagger-router-controller: botonPuntoPagos
    x-a127-apply: {}
    get:
      tags:
        - PuntoPago
      description: Recibe monto a recaudar por PuntoPagos
      # used as the method name of the controller
      operationId: puntoPagosValidar
      parameters:
        - name: idGenerado
          in: query
          description: Id de transacción generado por Punto Pagos
          required: true
          type: string
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/DespachaloResponse"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
...
definitions:
...
  Permission:
    type: object
    required:
      - typeUser
    properties:
      typeUser:
        description: Tipo de Usuario
        type: number
      createUser:
        description: Boolean permiso
        type: boolean
      editUser:
        description: Boolean permiso
        type: boolean
      editPermission:
        description: Boolean permiso
        type: boolean
      resetPassword:
        description: Boolean permiso
        type: boolean
  PermissionDelete:
    type: object
    required:
      - typeUser
    properties:
      typeUser:
        description: Tipo de Usuario
        type: number
...
  Notificacion:
    type: object
    required:
      - token
      - trx_id
      - medio_pago
      - monto
      - fecha_aprovacion
      - numero_tarjeta
      - num_cuotas
      - tipo_cuotas
      - valor_cuota
      - primer_vencimiento
      - numero_operacion
      - codigo_autorizacion
    properties:
      token:
        type: string
        description: Token
      trx_id:
        type: string
        description: Id de transacción
      medio_pago:
        type: string
        description: Medio de pago
      monto:
        type: string
        description: Monto
      fecha_aprovacion:
        type: string
        description: Fecha de aprovación
      numero_tarjeta:
        type: string
        description: Número de tarjeta
      num_cuotas:
        type: string
        description: Número de cuotas
      tipo_cuotas:
        type: string
        description: Número de cuotas
      valor_cuota:
        type: string
        description: Valor de cuotas
      primer_vencimiento:
        type: string
        description: Primer vencimiento
      numero_operacion:
        type: string
        description: Número de operación
      codigo_autorizacion:
        type: string
        description: Código de autorización
  DespachaloResponse:
    type: string
  ErrorResponse:
    type: string